1.增加教师任务定时发布功能 2.增加根据姓名生成头像的功能 3.优化如果教师任务已经发布，不可删除

+++ reading-cloud-book/src/main/java/cn/zealon/readingcloud/book/controller/TeacherTasksController.java
@@ -95,7 +95,14 @@ public class TeacherTasksController {
      */
     @DeleteMapping
     public ResponseEntity<Boolean> deleteById(Long id) {
-        return ResponseEntity.ok(this.teacherTasksService.deleteById(id));
+        TeacherTasks teacherTasks=this.teacherTasksService.queryById(id);
+        //判断如果任务已经发布，不可删除
+        if (teacherTasks != null) {
+            if (teacherTasks.getStatus() != 1) {
+                return ResponseEntity.ok(this.teacherTasksService.deleteById(id));
+            }
+        }
+        return ResponseEntity.ok(false);
     }
 
 }
